pyproject.toml
[project]
name = "template-production"
version = "0.1.0"
description = "Moative's production ready repository template"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "annotated-types==0.7.0",
    "anyio==3.7.1",
    "bcrypt>=4.3.0",
    "blinker==1.9.0",
    "cachelib>=0.13.0",
    "cachetools==5.5.2",
    "certifi==2025.4.26",
    "cfgv==3.4.0",
    "charset-normalizer==3.4.2",
    "click==8.2.1",
    "distlib==0.3.9",
    "distro==1.9.0",
    "elevenlabs>=2.3.0",
    "filelock==3.18.0",
    "filetype==1.2.0",
    "fastapi==0.104.1",
    "uvicorn[standard]==0.24.0",
    "python-multipart==0.0.6",
    "jinja2==3.1.6",
    "python-jose[cryptography]==3.3.0",
    "passlib[bcrypt]==1.7.4",
    "gevent>=25.5.1",
    "google-ai-generativelanguage==0.6.18",
    "google-api-core==2.25.0",
    "google-auth==2.40.2",
    "googleapis-common-protos==1.70.0",
    "grpcio==1.72.1",
    "grpcio-status==1.72.1",
    "gunicorn>=23.0.0",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "identify==2.6.12",
    "idna==3.10",
    "itsdangerous==2.2.0",
    "jiter==0.10.0",
    "jsonpatch==1.33",
    "jsonpointer==3.0.0",
    "langchain==0.3.25",
    "langchain-core==0.3.63",
    "langchain-google-genai==2.1.5",
    "langchain-openai==0.3.19",
    "langchain-text-splitters==0.3.8",
    "langgraph==0.4.8",
    "langgraph-checkpoint==2.0.26",
    "langgraph-checkpoint-sqlite>=2.0.10",
    "langgraph-prebuilt==0.2.2",
    "langgraph-sdk==0.1.70",
    "langsmith==0.3.44",
    "markupsafe==3.0.2",
    "nodeenv==1.9.1",
    "openai==1.83.0",
    "orjson==3.10.18",
    "ormsgpack==1.10.0",
    "packaging==24.2",
    "platformdirs==4.3.8",
    "pre-commit==4.2.0",
    "proto-plus==1.26.1",
    "protobuf==6.31.1",
    "pyasn1==0.6.1",
    "pyasn1-modules==0.4.2",
    "pydantic==2.7.4",
    "pydantic-core==2.18.4",
    "python-dotenv==1.1.0",
    "pyyaml==6.0.2",
    "regex==2024.11.6",
    "requests==2.32.3",
    "requests-toolbelt==1.0.0",
    "rsa==4.9.1",
    "ruff==0.11.12",
    "scipy>=1.15.3",
    "sniffio==1.3.1",
    "sounddevice>=0.5.2",
    "sqlalchemy==2.0.41",
    "tenacity==9.1.2",
    "tiktoken==0.9.0",
    "tqdm==4.67.1",
    "twilio>=9.6.2",
    "typing-extensions==4.14.0",
    "typing-inspection==0.4.1",
    "urllib3==2.4.0",
    "virtualenv==20.31.2",
    "werkzeug==3.1.3",
    "xxhash==3.5.0",
    "zstandard==0.23.0",
    "requests>=2.31.0",
    "defusedxml>=0.7.1",
    "google-maps-addressvalidation>=0.3.19",
    "pytz>=2025.2",
    "psutil>=7.0.0",
    "websockets>=12.0",
    "pyaudio>=0.2.14",
]

[dependency-groups]
dev = [
    "pre-commit>=4.0.1",
    "ruff>=0.8.1",
]
